refactor: simplify bootstrap for builtin extension model
- Replace 500-line auto-bootstrap with explicit bootstrap_openvk() function
- Move helper functions (tr, ovk_proc_strtr, etc.) and boot closure
  to ovk-init.php (loaded by ExtensionManager as init script)
- Remove manifest.yml (no longer needed)
- Update chandler_loader.php to use new bootstrap_openvk()
