+++ chandler_loader.php
@@ -6,5 +6,5 @@ declare(strict_types=1);
 namespace openvk;
 
 $_SERVER["HTTP_ACCEPT_LANGUAGE"] = false;
-$bootstrap = require(__DIR__ . "/../../../chandler/Bootstrap.php");
-$bootstrap->ignite(true);
+require __DIR__ . "/bootstrap.php";
+bootstrap_openvk(true);
