--- manifest.yml
@@ -1,6 +1,0 @@
-name: "OpenVK"
-description: "Not-yet-federated open source social network inspired by VK"
-author: "OpenVK Team"
-version: "0.1.0-tp3"
-
-init: "bootstrap.php"
